# src/main/resources/api/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
...
paths:
...
  "/bookings/{id}":
    get:
      tags:
        - Bookings
      summary: Получить бронирование по ID
      operationId: bookingsIdGet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/BookingResponse"
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '404':
          description: Бронирование не найдено
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
      security:
        - userAuth: []
    put:
      tags:
        - Bookings
      summary: Обновить бронирование
      operationId: bookingsIdPut
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/BookingUpdateRequest"
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/BookingResponse"
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '404':
          description: Бронирование не найдено
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
      security:
        - userAuth: []
    delete:
      tags:
        - Bookings
      summary: Отменить бронирование
      operationId: bookingsIdDelete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '204':
          description: Бронирование отменено
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '404':
          description: Бронирование не найдено
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
      security:
        - userAuth: []
  "/admin/bookings/{id}/confirm":
    put:
      tags:
        - Admin
      summary: Подтвердить бронирование (Admin)
      operationId: adminBookingsIdConfirm
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '204':
          description: Бронирование подтверждено
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '404':
          description: Бронирование не найдено
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
      security:
        - bearerAuth: []
...
  "/bookings":
    get:
      tags:
        - Bookings
      summary: Получить свои бронирования
      operationId: bookingsGet
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/BookingResponse"
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
      security:
        - userAuth: []
    post:
      tags:
        - Bookings
      summary: Создать бронирование
      operationId: bookingsPost
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CreateBookingRequest"
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/BookingResponse"
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '404':
          description: Бронирование не найдено
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
      security:
        - userAuth: []
  "/auth/register":
    post:
      tags:
        - Auth
      summary: Регистрация пользователя
      operationId: authRegisterPost
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/RegisterRequest"
        required: true
      responses:
        '201':
          description: Пользователь зарегистрирован
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/AuthResponse"
        '404':
          description: Ошибка валидации логина или пароля
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
  "/auth/login":
    post:
      tags:
        - Auth
      summary: Вход в систему
      operationId: authLoginPost
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/LoginRequest"
        required: true
      responses:
        '200':
          description: Успешный вход
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/AuthResponse"
        '401':
          description: Неверное имя пользователя или пароль
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
  "/admin/bookings":
    get:
      tags:
        - Admin
      summary: Получить все бронирования (Admin)
      operationId: adminBookingsGet
      parameters:
        - name: pageable
          in: query
          description: Параметры пагинации и сортировки
          required: true
          schema:
            "$ref": "#/components/schemas/Pageable"
      responses:
        '200':
          description: Страница с бронированиями
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/PageableBookingResponse"
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
      security:
        - bearerAuth: []
  "/admin/bookings/{id}":
    delete:
      tags:
        - Admin
      summary: Удалить бронирование (Admin)
      operationId: adminBookingsIdDelete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '204':
          description: Бронирование удалено
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '404':
          description: Бронирование не найдено
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
      security:
        - bearerAuth: []
components:
  schemas:
    ErrorDetail:
      type: object
      properties:
        field:
          type: string
          example: email
        message:
          type: string
          example: Должен быть валидным email адресом
      description: Детали ошибки (опционально)
    ErrorResponse:
      required:
        - code
        - level
        - message
      type: object
      properties:
        code:
          type: string
          description: Машинно-читаемый код ошибки
          example: VALIDATION_FAILED
        level:
          type: string
          description: Уровень серьезности
          example: error
          enum:
            - ErrorResponse.LevelEnum.INFO(value=info)
            - ErrorResponse.LevelEnum.WARNING(value=warning)
            - ErrorResponse.LevelEnum.ERROR(value=error)
        message:
          type: string
          description: Человеко-читаемое сообщение
          example: Некоторые поля заполнены неверно
        details:
          type: array
          description: Детали ошибки (опционально)
          items:
            "$ref": "#/components/schemas/ErrorDetail"
...
    BookingResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
        eventId:
          type: integer
          format: int32
        customerEmail:
          type: string
        ticketCount:
          type: integer
          format: int32
        bookingTime:
          type: string
          format: date-time
        expiryTime:
          type: string
          format: date-time
        confirmed:
          type: boolean
        timezone:
          type: object
          properties:
            id:
              type: string
            rules:
              type: object
              properties:
                fixedOffset:
                  type: boolean
                transitions:
                  type: array
                  items:
                    type: object
                    properties:
                      offsetBefore:
                        type: object
                        properties:
                          totalSeconds:
                            type: integer
                            format: int32
                          id:
                            type: string
                      offsetAfter:
                        type: object
                        properties:
                          totalSeconds:
                            type: integer
                            format: int32
                          id:
                            type: string
                      instant:
                        type: string
                        format: date-time
                      duration:
                        type: object
                        properties:
                          seconds:
                            type: integer
                            format: int64
                          zero:
                            type: boolean
                          nano:
                            type: integer
                            format: int32
                          negative:
                            type: boolean
                          units:
                            type: array
                            items:
                              type: object
                              properties:
                                durationEstimated:
                                  type: boolean
                                timeBased:
                                  type: boolean
                                dateBased:
                                  type: boolean
                      gap:
                        type: boolean
                      dateTimeAfter:
                        type: string
                        format: date-time
                      dateTimeBefore:
                        type: string
                        format: date-time
                      overlap:
                        type: boolean
                transitionRules:
                  type: array
                  items:
                    type: object
                    properties:
                      month:
                        type: string
                        enum:
                          - JANUARY
                          - FEBRUARY
                          - MARCH
                          - APRIL
                          - MAY
                          - JUNE
                          - JULY
                          - AUGUST
                          - SEPTEMBER
                          - OCTOBER
                          - NOVEMBER
                          - DECEMBER
                      timeDefinition:
                        type: string
                        enum:
                          - UTC
                          - WALL
                          - STANDARD
                      standardOffset:
                        type: object
                        properties:
                          totalSeconds:
                            type: integer
                            format: int32
                          id:
                            type: string
                      offsetBefore:
                        type: object
                        properties:
                          totalSeconds:
                            type: integer
                            format: int32
                          id:
                            type: string
                      offsetAfter:
                        type: object
                        properties:
                          totalSeconds:
                            type: integer
                            format: int32
                          id:
                            type: string
                      dayOfWeek:
                        type: string
                        enum:
                          - MONDAY
                          - TUESDAY
                          - WEDNESDAY
                          - THURSDAY
                          - FRIDAY
                          - SATURDAY
                          - SUNDAY
                      dayOfMonthIndicator:
                        type: integer
                        format: int32
                      localTime:
                        "$ref": "#/components/schemas/LocalTime"
                      midnightEndOfDay:
                        type: boolean
    LocalTime:
      type: object
      properties:
        hour:
          type: integer
          format: int32
        minute:
          type: integer
          format: int32
        second:
          type: integer
          format: int32
        nano:
          type: integer
          format: int32
    BookingUpdateRequest:
      type: object
      properties:
        ticketCount:
          minimum: 1
          type: integer
          format: int32
...
    CreateBookingRequest:
      required:
        - eventId
        - ticketCount
      type: object
      properties:
        eventId:
          type: integer
          format: int32
        ticketCount:
          minimum: 1
          type: integer
          format: int32
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        role:
          type: string
          enum:
            - USER
            - ADMIN
    RegisterRequest:
      required:
        - email
        - password
      type: object
      properties:
        email:
          type: string
        password:
          maxLength: 2147483647
          minLength: 8
          type: string
    LoginRequest:
      required:
        - email
        - password
      type: object
      properties:
        email:
          type: string
        password:
          type: string
...
    PageableObject:
      type: object
      properties:
        paged:
          type: boolean
        pageSize:
          type: integer
          format: int32
        pageNumber:
          type: integer
          format: int32
        unpaged:
          type: boolean
        offset:
          type: integer
          format: int64
        sort:
          type: array
          items:
            "$ref": "#/components/schemas/SortObject"
    SortObject:
      type: object
      properties:
        direction:
          type: string
        nullHandling:
          type: string
        ascending:
          type: boolean
        property:
          type: string
        ignoreCase:
          type: boolean
    Pageable:
      type: object
      properties:
        page:
          minimum: 0
          type: integer
          format: int32
        size:
          minimum: 1
          type: integer
          format: int32
        sort:
          type: array
          items:
            type: string
    PageableBookingResponse:
      type: object
      properties:
        content:
          type: array
          items:
            "$ref": "#/components/schemas/BookingResponse"
        pageable:
          "$ref": "#/components/schemas/PageableObject"
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        last:
          type: boolean
        size:
          type: integer
          format: int32
        number:
          type: integer
          format: int32
        sort:
          type: array
          items:
            "$ref": "#/components/schemas/SortObject"
        first:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        empty:
          type: boolean
      description: Страница с бронированиями